backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.13
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
pydantic[email]>=2.0
email-validator>=2.0.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
jwt
python-jose[cryptography]
requests
openai
dotenv
langchain 
python-dotenv
langchain-community
langchain-openai
# googleapiclient
google-api-python-client
google-auth-oauthlib
# openai==1.3.9
notion-client
